# Improve performance of scoring and video creation