--- a/package.json
+++ b/package.json
@@ -28,7 +28,8 @@
     "express": "^4.18.2",
     "nanoid": "^4.0.1",
     "puppeteer": "^19.6.3",
-    "puppeteer-mass-screenshots": "^1.0.15"
+    "puppeteer-mass-screenshots": "^1.0.15",
+    "uuid": "^9.0.0"
   },
   "devDependencies": {
     "@types/airtable": "^0.10.1",
@@ -36,6 +37,7 @@
     "@types/dotenv": "^8.2.0",
     "@types/express": "^4.17.17",
     "@types/nanoid": "^3.0.0",
+    "@types/uuid": "^9.0.1",
     "ts-node": "^10.9.1",
     "typescript": "^5.0.2"
   }