This PR does a few things.

- Support for utilizing the `ticksPerSecond` and `drawModulo` query parameters that will be available when [this pr](https://github.com/hackclub/sinerider/pull/303) lands.
- Reduces the amount of overall scoring processing time for the default level (Welcome / HELLO_WORLD) from ~30 seconds to <10 seconds (300% increase).  This is done by increasing the ticks per second to 1000 (compared to a normal 30).
- Adds support for timing out scores (fails to score things when they take longer than 30s to simulate).  This also leverages the time scaling, so we will not block a server thread for 30 seconds in this case.
- Reduces the resolution and encoding settings of both the puppeteer + ffmpeg, reducing ultimate video fidelity.  This can be carefully adjusted and even modified in realtime in the service depending on load, we should do some experimentation on what levers we want to pull how much.
- Time-scales the video that we produce, which again can be controlled carefully with future experiments.  We can render videos that in realtime take 15 seconds to simulate, but show a brief synopsis in <5 sec (fast-forwarded)
- Puts image and video content into a temporary directory which is deleted afterwards, cleaning up all evidence.